avance inicio de sesión

+++ backend/src/main/java/com/libruarys/repository/UsuarioRepository.java
@@ -2,9 +2,8 @@ package com.libruarys.repository;
 
 import com.libruarys.model.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional; // <--- ESTA LÍNEA ES LA QUE FALTA
+import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
-    // Esto permite buscar al usuario para validar su contraseña
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Optional<Usuario> findByCorreo(String correo);
 }
